+++ extensions/thesaurus-rex/src/look-up-word.tsx
@@ -1,0 +1,5 @@
+import LookupView from "./lookup-view";
+
+export default function LookUpWord() {
+  return <LookupView kinds={["synonyms", "antonyms", "definitions"]} />;
+}